Se añaden los bloques alias, funcion y niveles.
git-svn-id: file:///cygdrive/c/CTTC/backup/redyc/svn/Colossus/trunk@238 8e1883df-7ee1-ef4f-9e60-6d04f2a2cf0e

+++ src/win32/debug.c
@@ -1,5 +1,5 @@
 /*
- * $Id: debug.c,v 1.7 2005-11-01 15:51:01 Trocotronic Exp $ 
+ * $Id: debug.c,v 1.8 2005-12-04 14:09:25 Trocotronic Exp $ 
  */
 
 #include "struct.h"
@@ -9,7 +9,7 @@
 
 BOOL CALLBACK SymEnumerateModulesProc64(PCSTR ModuleName, DWORD64 BaseOfDll, PVOID UserContext)
 {
-    SymUnloadModule((HANDLE)UserContext, BaseOfDll);
+    SymUnloadModule((HANDLE)UserContext, (DWORD)BaseOfDll);
     return TRUE;
 }
 __inline char *StackTrace(EXCEPTION_POINTERS *e)
